Add the linalg module
