+++ src/lib.rs
@@ -5,5 +5,6 @@ pub mod algorithm;
 pub mod collections;
 pub mod exporters;
 pub mod importers;
+pub mod linalg;
 pub mod protocol_traits;
 pub mod types;
